Move faker provider
